+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1 (not yet released)
 
 - updated `GroovyXmlTransform` to understand new XML element `<host-excludes>`
+- added `OnlineOptions.bootTimeout`
 
 ## 1.0.0 [The Ez|Ra Edition](https://en.wikipedia.org/wiki/Embassytown) (2016-03-15)